feat: Implementacao completa do fluxo de cadastro de empresa com corr…

+++ src/main/java/com/anatonelly/freteexpress/repository/FreteRepository.java
@@ -1,7 +1,7 @@
 package com.anatonelly.freteexpress.repository;
 
-import com.anatonelly.freteexpress.model.EmpresaCliente;
 import com.anatonelly.freteexpress.model.Frete;
+import com.anatonelly.freteexpress.model.EmpresaCliente;
 import com.anatonelly.freteexpress.model.Motorista;
 import com.anatonelly.freteexpress.enums.StatusFrete;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -10,8 +10,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface FreteRepository extends JpaRepository<Frete, Long> {
-
+public interface FreteRepository extends JpaRepository<Frete, Integer> { // <<<<< CORRIGIDO AQUI: Long para Integer
     List<Frete> findByMotoristaSolicitanteAndStatus(Motorista motorista, StatusFrete status);
     List<Frete> findByEmpresaClienteOrderByPrazoEntregaDesc(EmpresaCliente empresaCliente);
 }
