+++ src/main/java/com/anatonelly/freteexpress/service/FreteService.java
@@ -3,7 +3,7 @@ package com.anatonelly.freteexpress.service;
 import com.anatonelly.freteexpress.model.EmpresaCliente;
 import com.anatonelly.freteexpress.model.Frete;
 import com.anatonelly.freteexpress.model.Motorista;
-import com.anatonelly.freteexpress.enums.StatusFrete;
+import com.anatonelly.freteexpress.enums.StatusFrete; // Certifique-se que este enum existe e tem os valores
 import com.anatonelly.freteexpress.repository.FreteRepository;
 import com.anatonelly.freteexpress.repository.MotoristaRepository;
 import org.springframework.stereotype.Service;
@@ -26,7 +26,9 @@ public class FreteService {
         return freteRepository.findAll();
     }
 
-    public Optional<Frete> buscarPorId(Long id) {
+    // O ID da entidade Frete deve ser Long, se você usou Long no JpaRepository
+    // Se o ID for Integer, ajuste aqui também
+    public Optional<Frete> buscarPorId(Integer id) { // Alterado para Integer
         return freteRepository.findById(id);
     }
 
@@ -34,7 +36,7 @@ public class FreteService {
         return freteRepository.save(frete);
     }
 
-    public boolean deletar(Long id) {
+    public boolean deletar(Integer id) { // Alterado para Integer
         if (freteRepository.existsById(id)) {
             freteRepository.deleteById(id);
             return true;
@@ -42,19 +44,22 @@ public class FreteService {
         return false;
     }
 
-    public Optional<Frete> atualizar(Long id, Frete freteAtualizado) {
+    // O ID da entidade Frete deve ser Long, se você usou Long no JpaRepository
+    // Se o ID for Integer, ajuste aqui também
+    public Optional<Frete> atualizar(Integer id, Frete freteAtualizado) { // Alterado para Integer
         return freteRepository.findById(id).map(freteExistente -> {
-            freteExistente.setOrigem(freteAtualizado.getOrigem());
-            freteExistente.setDestino(freteAtualizado.getDestino());
-            freteExistente.setTipoCarga(freteAtualizado.getTipoCarga());
-            freteExistente.setPeso(freteAtualizado.getPeso());
-            freteExistente.setDimensoes(freteAtualizado.getDimensoes());
-            freteExistente.setPrazoEntrega(freteAtualizado.getPrazoEntrega());
-            freteExistente.setDescricaoCarga(freteAtualizado.getDescricaoCarga());
-            freteExistente.setValorFrete(freteAtualizado.getValorFrete());
-            freteExistente.setStatus(freteAtualizado.getStatus());
-            freteExistente.setEmpresaCliente(freteAtualizado.getEmpresaCliente());
-            freteExistente.setMotoristaSolicitante(freteAtualizado.getMotoristaSolicitante());
+            // Certifique-se que os getters/setters abaixo estão na sua classe Frete.java
+            freteExistente.setOrigem(freteAtualizado.getOrigem()); // Exemplo, verificar se existe na classe Frete
+            freteExistente.setDestino(freteAtualizado.getDestino()); // Exemplo
+            freteExistente.setTipoCarga(freteAtualizado.getTipoCarga()); // Exemplo
+            freteExistente.setPeso(freteAtualizado.getPeso()); // Exemplo
+            freteExistente.setDimensoes(freteAtualizado.getDimensoes()); // Exemplo
+            freteExistente.setPrazoEntrega(freteAtualizado.getPrazoEntrega()); // Exemplo
+            freteExistente.setDescricaoCarga(freteAtualizado.getDescricaoCarga()); // Exemplo
+            freteExistente.setValorFrete(freteAtualizado.getValorFrete()); // Exemplo
+            freteExistente.setStatus(freteAtualizado.getStatus()); // Exemplo
+            freteExistente.setEmpresaCliente(freteAtualizado.getEmpresaCliente()); // Exemplo
+            freteExistente.setMotoristaSolicitante(freteAtualizado.getMotoristaSolicitante()); // Exemplo
             return freteRepository.save(freteExistente);
         });
     }
@@ -67,15 +72,15 @@ public class FreteService {
         return freteRepository.findByEmpresaClienteOrderByPrazoEntregaDesc(empresaCliente);
     }
 
-    public boolean solicitarFrete(Long freteId, Long motoristaId) {
+    public boolean solicitarFrete(Integer freteId, Integer motoristaId) { // <<<<< CORRIGIDO AQUI: Long para Integer
         Optional<Frete> freteOpt = freteRepository.findById(freteId);
         Optional<Motorista> motoristaOpt = motoristaRepository.findById(motoristaId);
 
         if (freteOpt.isPresent() && motoristaOpt.isPresent()) {
             Frete frete = freteOpt.get();
-            if (frete.getStatus() == StatusFrete.PENDENTE) {
-                frete.setStatus(StatusFrete.SOLICITADO);
-                frete.setMotoristaSolicitante(motoristaOpt.get());
+            if (frete.getStatus() == StatusFrete.PENDENTE) { // Certifique-se que StatusFrete.PENDENTE existe
+                frete.setStatus(StatusFrete.SOLICITADO); // Certifique-se que StatusFrete.SOLICITADO existe
+                frete.setMotoristaSolicitante(motoristaOpt.get()); // Certifique-se que este setter existe em Frete
                 freteRepository.save(frete);
                 return true;
             }
